GATHERER OVERHAUL: core.gatherer -> utils.gather, demoted from core class to utility package utils.gather functions are being stripped down to bare essentials, no more reliance on internal path attributes etc. manager.gather -> manager.gather_from_dataset, used to get internal data from a dataset the underlying motivation here is to make data gathering much more explicit because it is currently very stupidly implicit and difficult to track/manipulate
